+++ R/DE_bind_CA.R
@@ -5,8 +5,8 @@
 #' @param atac.peak.res Dataframe contains peak annotation results of ATAC-seq data.
 #' @param peak.anno.key Peak location, chosen from "Promoter", "5' UTR", "3' UTR", "Exon", "Intron", "Downstream", "Distal Intergenic","All". Default: "Promoter".
 #' @param signif Significance criterion. For DESeq2 results, can be chosen from padj, pvalue. For edgeR results, can be chosen from FDR, PValue. Default: padj.
-#' @param signif.threashold Significance threashold to get differentially expressed genes. Default: 0.05.
-#' @param l2fc.threashold Log2 fold change threashold to get differentially expressed genes. Default: 1.
+#' @param signif.threshold Significance threshold to get differentially expressed genes. Default: 0.05.
+#' @param l2fc.threshold Log2 fold change threshold to get differentially expressed genes. Default: 1.
 #' @param label.key Which column to use as label. Default: NULL (use rownames of \code{deres}).
 #' @param merge.key The columns used for merging, chosen from geneId, ENSEMBL, SYMBOL. Default: geneId.
 #'
@@ -66,7 +66,7 @@
 #'   peak.anno.key = "Promoter", merge.key = "SYMBOL"
 #' )
 DEbCA <- function(de.res, chip.peak.res, atac.peak.res, peak.anno.key = c("Promoter", "5' UTR", "3' UTR", "Exon", "Intron", "Downstream", "Distal Intergenic", "All"),
-                  signif = "padj", signif.threashold = 0.05, l2fc.threashold = 1, label.key = NULL, merge.key = c("geneId", "ENSEMBL", "SYMBOL")) {
+                  signif = "padj", signif.threshold = 0.05, l2fc.threshold = 1, label.key = NULL, merge.key = c("geneId", "ENSEMBL", "SYMBOL")) {
   # check parameters
   peak.anno.key <- match.arg(arg = peak.anno.key)
   merge.key <- match.arg(arg = merge.key)
@@ -88,8 +88,8 @@ DEbCA <- function(de.res, chip.peak.res, atac.peak.res, peak.anno.key = c("Promo
 
   # Integrate RNA-seq and ChIP-seq results
   debchip.res <- DEbPeak(
-    de.res = de.res, peak.res = chip.peak.df, peak.anno.key = peak.anno.key, signif = signif, signif.threashold = signif.threashold,
-    l2fc.threashold = l2fc.threashold, label.key = label.key, merge.key = merge.key
+    de.res = de.res, peak.res = chip.peak.df, peak.anno.key = peak.anno.key, signif = signif, signif.threshold = signif.threshold,
+    l2fc.threshold = l2fc.threshold, label.key = label.key, merge.key = merge.key
   )
   # Integrate above results with ATAC-seq results
   # change debchip results name